requirements.txt
numba
numpy
pandas
matplotlib
pillow
hilbertcurve
cupy-cuda12x
